seconds2str() moved to utils
